Make `dock-dbus-desktop-file-name' dynamic
Having `dock--dbus-desktop-id' to be computed when `dock-mode' is
activated produces some counterintuitive behavior. For example this code
doesn't work:

    (require 'dock)
    (dock-mode +1)
    (let ((dock-dbus-desktop-file-name "org-mode"))
      (dock-badge 53))

Even though it is a very edge case, I don't think it makes much sense to
store a computed value and disconnect the custom variable from the
functions which use it.